move some of internal/codeintel/stores/lsifstore => lib/codeintel/semantic
