+++ enterprise/lib/codeintel/semantic/hash_test.go
@@ -1,4 +1,4 @@
-package lsifstore
+package semantic
 
 import (
 	"testing"
