+++ enterprise/lib/codeintel/semantic/util_test.go
@@ -1,4 +1,4 @@
-package lsifstore
+package semantic
 
 import (
 	"strconv"
